requirement.txt
python-dotenv==0.21.0
psycopg2==2.9.5
Faker==15.3.4
rich==12.6.0
uuid==1.30